majit rbuilder: relocate builder runtime to pyre-object; bind ll_append residual fnaddr shims
Move `StringBuilderBox`/`StringPieceBox` and the `rbuilder_runtime` module
(`ll_new`, `ll_grow_by`, `ll_grow_and_append`, `ll_append`, `ll_append_char`,
`ll_getlength`, `ll_shrink_final`, `ll_fold_pieces`, `ll_build`, plus the #52
semantics tests) from `pyre-jit::eval` to `pyre_object::rbuilder`. Move the raw
low-level-string helpers `bh_lowlevel_chars_offset`, `bh_read_lowlevel_string`,
`bh_write_lowlevel_char` to `pyre_object::lowlevel_string` beside their existing
`bh_alloc`/`bh_free`/`bh_len` siblings. The relocated crate is reachable from
`pyre-interpreter`, where the fnaddr registry lives; `pyre-jit::eval` was not.

Add `pub extern "C"` shims `jit_ll_append_res0(builder, ll_str)` and
`jit_ll_append_res_slice(builder, ll_str, start, end)` delegating to `ll_append`
with `item_size = 1`, and register them in `jit_trace_fnaddrs` under the
un-prefixed leaf keys `ll_append_res0` / `ll_append_res_slice`. With the fnaddr
bound, `guess_call_kind`'s `dont_look_inside` residual arm (the append helper
recognizer) resolves these callsites.

`build_gc`'s builder/piece tid registration now reads the
`pyre_object::rbuilder` offset/size consts instead of `size_of`/`offset_of` on
the moved structs, so `pyre-jit::eval` no longer references the struct types.
